rename: components 폴더 경로 변경
