+++ src/app/page.tsx
@@ -1,7 +1,7 @@
-import RemoteControl from '@/components/features/Common/RemoteControl/RemoteControl';
-import Card from '@/components/features/Items/Card/Card';
-import Header from '@/components/widgets/Header/Header';
-import SearchToolbar from '@/components/widgets/SearchToolbar/SearchToolbar';
+import RemoteControl from '@/commons/components/features/Common/RemoteControl/RemoteControl';
+import Card from '@/commons/components/features/Items/Card/Card';
+import Header from '@/commons/components/widgets/Header/Header';
+import SearchToolbar from '@/commons/components/widgets/SearchToolbar/SearchToolbar';
 import * as styles from '@/commons/styles/rootStyles.css';
 
 export default function RootPage() {
