+++ src/commons/components/atoms/Icons/MagnifyingGlass/MagnifyingGlass.tsx
