+++ src/commons/components/atoms/Icons/Sun/Sun.tsx
