+++ src/commons/components/features/Common/Slider/Slider.css.ts
