+++ src/commons/components/features/Common/Slider/Slider.tsx
