+++ src/commons/components/widgets/SearchToolbar/SearchToolbar.css.ts
